+++ x/evm/types/tracer.go
@@ -6,6 +6,8 @@ import (
 	"os"
 	"time"
 
+	"github.com/ethereum/go-ethereum/eth/tracers/logger"
+
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/vm"
@@ -21,22 +23,22 @@ const (
 
 // NewTracer creates a new Logger tracer to collect execution traces from an
 // EVM transaction.
-func NewTracer(tracer string, msg core.Message, cfg *params.ChainConfig, height int64) vm.Tracer {
+func NewTracer(tracer string, msg core.Message, cfg *params.ChainConfig, height int64) vm.EVMLogger {
 	// TODO: enable additional log configuration
-	logCfg := &vm.LogConfig{
+	logCfg := &logger.Config{
 		Debug: true,
 	}
 
 	switch tracer {
 	case TracerAccessList:
-		precompiles := vm.ActivePrecompiles(cfg.Rules(big.NewInt(height)))
-		return vm.NewAccessListTracer(msg.AccessList(), msg.From(), *msg.To(), precompiles)
+		preCompiles := vm.ActivePrecompiles(cfg.Rules(big.NewInt(height)))
+		return logger.NewAccessListTracer(msg.AccessList(), msg.From(), *msg.To(), preCompiles)
 	case TracerJSON:
-		return vm.NewJSONLogger(logCfg, os.Stderr)
+		return logger.NewJSONLogger(logCfg, os.Stderr)
 	case TracerMarkdown:
-		return vm.NewMarkdownLogger(logCfg, os.Stdout) // TODO: Stderr ?
+		return logger.NewMarkdownLogger(logCfg, os.Stdout) // TODO: Stderr ?
 	case TracerStruct:
-		return vm.NewStructLogger(logCfg)
+		return logger.NewStructLogger(logCfg)
 	default:
 		return NewNoOpTracer()
 	}
@@ -79,7 +81,7 @@ type StructLogRes struct {
 }
 
 // FormatLogs formats EVM returned structured logs for json output
-func FormatLogs(logs []vm.StructLog) []StructLogRes {
+func FormatLogs(logs []logger.StructLog) []StructLogRes {
 	formatted := make([]StructLogRes, len(logs))
 	for index, trace := range logs {
 		formatted[index] = StructLogRes{
@@ -123,7 +125,7 @@ func FormatLogs(logs []vm.StructLog) []StructLogRes {
 	return formatted
 }
 
-var _ vm.Tracer = &NoOpTracer{}
+var _ vm.EVMLogger = &NoOpTracer{}
 
 // NoOpTracer is an empty implementation of vm.Tracer interface
 type NoOpTracer struct{}
@@ -134,60 +136,23 @@ func NewNoOpTracer() *NoOpTracer {
 }
 
 // CaptureStart implements vm.Tracer interface
-func (dt NoOpTracer) CaptureStart(
-	env *vm.EVM,
-	from, to common.Address,
-	create bool,
-	input []byte,
-	gas uint64,
-	value *big.Int,
-) {
+func (dt NoOpTracer) CaptureStart(env *vm.EVM, from common.Address, to common.Address, create bool, input []byte, gas uint64, value *big.Int) {
 }
 
+// CaptureState implements vm.Tracer interface
+func (dt NoOpTracer) CaptureState(pc uint64, op vm.OpCode, gas, cost uint64, scope *vm.ScopeContext, rData []byte, depth int, err error) {
+}
+
+// CaptureFault implements vm.Tracer interface
+func (dt NoOpTracer) CaptureFault(pc uint64, op vm.OpCode, gas, cost uint64, scope *vm.ScopeContext, depth int, err error) {
+}
+
+// CaptureEnd implements vm.Tracer interface
+func (dt NoOpTracer) CaptureEnd(output []byte, gasUsed uint64, tm time.Duration, err error) {}
+
 // CaptureEnter implements vm.Tracer interface
-func (dt NoOpTracer) CaptureEnter(
-	typ vm.OpCode,
-	from common.Address,
-	to common.Address,
-	input []byte,
-	gas uint64,
-	value *big.Int,
-) {
+func (dt NoOpTracer) CaptureEnter(typ vm.OpCode, from common.Address, to common.Address, input []byte, gas uint64, value *big.Int) {
 }
 
 // CaptureExit implements vm.Tracer interface
 func (dt NoOpTracer) CaptureExit(output []byte, gasUsed uint64, err error) {}
-
-// CaptureState implements vm.Tracer interface
-func (dt NoOpTracer) CaptureState(
-	env *vm.EVM,
-	pc uint64,
-	op vm.OpCode,
-	gas, cost uint64,
-	scope *vm.ScopeContext,
-	rData []byte,
-	depth int,
-	err error,
-) {
-}
-
-// CaptureFault implements vm.Tracer interface
-func (dt NoOpTracer) CaptureFault(
-	env *vm.EVM,
-	pc uint64,
-	op vm.OpCode,
-	gas, cost uint64,
-	scope *vm.ScopeContext,
-	depth int,
-	err error,
-) {
-}
-
-// CaptureEnd implements vm.Tracer interface
-func (dt NoOpTracer) CaptureEnd(
-	output []byte,
-	gasUsed uint64,
-	t time.Duration,
-	err error,
-) {
-}