+++ x/evm/types/tracer_test.go
@@ -2,10 +2,10 @@ package types
 
 import (
 	"fmt"
+	"github.com/ethereum/go-ethereum/eth/tracers/logger"
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/holiman/uint256"
 	"github.com/stretchr/testify/require"
 )
@@ -17,17 +17,17 @@ func TestFormatLogs(t *testing.T) {
 
 	testCases := []struct {
 		name string
-		logs []vm.StructLog
+		logs []logger.StructLog
 		exp  []StructLogRes
 	}{
 		{
 			"empty logs",
-			[]vm.StructLog{},
+			[]logger.StructLog{},
 			[]StructLogRes{},
 		},
 		{
 			"non-empty stack",
-			[]vm.StructLog{
+			[]logger.StructLog{
 				{
 					Stack: zeroUint256,
 				},
@@ -42,7 +42,7 @@ func TestFormatLogs(t *testing.T) {
 		},
 		{
 			"non-empty memory",
-			[]vm.StructLog{
+			[]logger.StructLog{
 				{
 					Memory: zeroByte,
 				},
@@ -57,7 +57,7 @@ func TestFormatLogs(t *testing.T) {
 		},
 		{
 			"non-empty storage",
-			[]vm.StructLog{
+			[]logger.StructLog{
 				{
 					Storage: make(map[common.Hash]common.Hash),
 				},